requirements.txt
fastapi
uvicorn
supabase
python-dotenv
numpy
scikit-learn